requirements.txt
django-cors-headers==4.2.0
django-rest-auth==0.9.5
Django==4.2.5
django-allauth==0.52.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0